Adaptive 15

+++ src/components/Popup/index.jsx
@@ -1,10 +1,16 @@
 import React from "react";
 import styles from "./Popup.module.scss";
+import Modal from "react-modal";
 
 const Popup = ({ setIsPopup, isPopup, setIsOpenModal, setIsOpenNewPost }) => {
   const outMyAccount = () => {
-    localStorage.clear();
-    window.location.reload();
+    if (window.confirm("Вы действительно хотите выйти?")) {
+      localStorage.clear();
+      window.location.reload();
+    }
+  };
+  const closeModal = () => {
+    setIsPopup(false);
   };
 
   return (
@@ -19,44 +25,61 @@ const Popup = ({ setIsPopup, isPopup, setIsOpenModal, setIsOpenNewPost }) => {
         </svg>
       </div>
       {isPopup ? (
-        <div className={styles.popup__open}>
-          <p onClick={() => setIsOpenNewPost(true)}>
-            <div width={30}>
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="rgba(92,92,232,1)"
-              >
-                <path d="M4 3H20C20.5523 3 21 3.44772 21 4V20C21 20.5523 20.5523 21 20 21H4C3.44772 21 3 20.5523 3 20V4C3 3.44772 3.44772 3 4 3ZM5 5V19H19V5H5ZM11 11V7H13V11H17V13H13V17H11V13H7V11H11Z"></path>
-              </svg>
-            </div>
-            Добавить пост
-          </p>
-          <p onClick={() => setIsOpenModal(true)}>
-            <div width={30}>
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="rgba(92,92,232,1)"
-              >
-                <path d="M12.8995 6.85453L17.1421 11.0972L7.24264 20.9967H3V16.754L12.8995 6.85453ZM14.3137 5.44032L16.435 3.319C16.8256 2.92848 17.4587 2.92848 17.8492 3.319L20.6777 6.14743C21.0682 6.53795 21.0682 7.17112 20.6777 7.56164L18.5563 9.68296L14.3137 5.44032Z"></path>
-              </svg>
-            </div>
-            Редактировать профиль
-          </p>
-          <p onClick={outMyAccount}>
-            <div width={30}>
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="rgba(92,92,232,1)"
-              >
-                <path d="M5 22C4.44772 22 4 21.5523 4 21V3C4 2.44772 4.44772 2 5 2H19C19.5523 2 20 2.44772 20 3V6H18V4H6V20H18V18H20V21C20 21.5523 19.5523 22 19 22H5ZM18 16V13H11V11H18V8L23 12L18 16Z"></path>
-              </svg>
-            </div>
-            Выйти из аккаунта
-          </p>
-        </div>
+        <Modal
+          className="popupModal"
+          isOpen={isPopup}
+          onRequestClose={closeModal}
+          ariaHideApp={false}
+          style={{
+            overlay: {
+              backgroundColor: "rgba(0, 0, 0, 0.75)",
+              zIndex: "100",
+            },
+            content: {
+              border: "none",
+              outline: "none",
+            },
+          }}
+        >
+          <div className={styles.popup__open}>
+            <p onClick={() => setIsOpenNewPost(true)}>
+              <div width={30}>
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 24 24"
+                  fill="rgba(92,92,232,1)"
+                >
+                  <path d="M4 3H20C20.5523 3 21 3.44772 21 4V20C21 20.5523 20.5523 21 20 21H4C3.44772 21 3 20.5523 3 20V4C3 3.44772 3.44772 3 4 3ZM5 5V19H19V5H5ZM11 11V7H13V11H17V13H13V17H11V13H7V11H11Z"></path>
+                </svg>
+              </div>
+              Добавить пост
+            </p>
+            <p onClick={() => setIsOpenModal(true)}>
+              <div width={30}>
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 24 24"
+                  fill="rgba(92,92,232,1)"
+                >
+                  <path d="M12.8995 6.85453L17.1421 11.0972L7.24264 20.9967H3V16.754L12.8995 6.85453ZM14.3137 5.44032L16.435 3.319C16.8256 2.92848 17.4587 2.92848 17.8492 3.319L20.6777 6.14743C21.0682 6.53795 21.0682 7.17112 20.6777 7.56164L18.5563 9.68296L14.3137 5.44032Z"></path>
+                </svg>
+              </div>
+              Редактировать профиль
+            </p>
+            <p onClick={outMyAccount}>
+              <div width={30}>
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 24 24"
+                  fill="rgba(92,92,232,1)"
+                >
+                  <path d="M5 22C4.44772 22 4 21.5523 4 21V3C4 2.44772 4.44772 2 5 2H19C19.5523 2 20 2.44772 20 3V6H18V4H6V20H18V18H20V21C20 21.5523 19.5523 22 19 22H5ZM18 16V13H11V11H18V8L23 12L18 16Z"></path>
+                </svg>
+              </div>
+              Выйти из аккаунта
+            </p>
+          </div>
+        </Modal>
       ) : (
         ""
       )}
